requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
flasgger
boto3
gunicorn
python-dotenv
pymysql
cryptography